hooking up category to delete
