+++ src/components/AdminCategory/AdminCategory.jsx
@@ -15,6 +15,7 @@ import {
   TableContainer,
   TablePagination,
 } from "@mui/material";
+import AdminCategoryRow from "./AdminCategoryRow";
 
 function AdminCategory() {
   const dispatch = useDispatch();
@@ -26,13 +27,7 @@ function AdminCategory() {
     (store) => store.adminReducer.adminCategoriesReducer
   );
 
-  // Defining what to do when edit or delete are pressed:
-  const handleEdit = () => {
-    console.log("Edit");
-  };
-  const handleDelete = () => {
-    console.log("Delete");
-  };
+  
 
   // The below 2 functions allow there to be multiple pages on the table.
   const handleChangePage = (event, newPage) => {
@@ -60,29 +55,15 @@ function AdminCategory() {
                 <TableCell align="center">Edit/Delete</TableCell>
               </TableRow>
             </TableHead>
-            <TableBody>
+           
               {catagories
                 .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                 .map((row) => (
-                  <TableRow>
-                    <TableCell component="th" scope="row" align="center">
-                      {row.name}
-                    </TableCell>
-                    <TableCell align="center">
-                      <Button
-                        sx={{ mr: 1 }}
-                        variant="contained"
-                        onClick={handleEdit}
-                      >
-                        Edit
-                      </Button>
-                      <Button variant="contained" onClick={handleDelete}>
-                        Delete
-                      </Button>
-                    </TableCell>
-                  </TableRow>
+                  <TableBody key={row.id}>
+                  <AdminCategoryRow row={row} />
+                  </TableBody>
                 ))}
-            </TableBody>
+           
           </Table>
         </TableContainer>
         <TablePagination
