+++ src/redux/sagas/categories.saga.js
@@ -19,7 +19,7 @@ function* fetchCategories() {
 }
 
 //CATEGORIES POST ROUTE
-function* addCategory() {
+function* addCategory(action) {
   // Get all for map category detail 
   try {
     //POST request to categories router
@@ -36,7 +36,7 @@ function* addCategory() {
 }
 
 //CATEGORIES PUT ROUTE
-function* updateCategory() {
+function* updateCategory(action) {
     try {
       //UPDATE request sent to categories.router based on ID
       const response = yield axios.put(`/api/categories/${action.payload}`);
@@ -52,7 +52,7 @@ function* updateCategory() {
   }
 
 //CATEGORIES DELETE ROUTE
-function* deleteCategory() {
+function* deleteCategory(action) {
     try {
       //DELETE request sent to categories.router based on ID
       const response = yield axios.delete(`/api/categories/${action.payload}`);
